Move tests in a separate file
